% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_sort_agg_column}
\alias{add_sort_agg_column}
\title{Add sort aggregator to the dataframe}
\usage{
add_sort_agg_column(df, sort_column_name, respect_grouping = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{sort_column_name}{(string) Name of the column to be used for sorting}

\item{respect_grouping}{(flag, default: TRUE) Whether the sorting should
honour group structure in the input dataframe (if any)}
}
\value{
Dataframe with new column added: '.quantile'
}
\description{
The added column is always named '.quantile'. This acts as an
  aggregator for downstream functions
}
