% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoplot.data.frame}
\alias{autoplot.data.frame}
\title{autplot method for dataframe}
\usage{
\method{autoplot}{data.frame}(object, sort_column_name, respect_grouping = TRUE)
}
\arguments{
\item{object}{data.frame}

\item{sort_column_name}{(string) Name of the column to be used for sorting}

\item{respect_grouping}{(flag, default: TRUE) Whether the sorting should
honour group structure in the input dataframe (if any)}
}
\value{
patchwork object
}
\description{
autplot method for dataframe
}
