% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dfvis.R
\docType{package}
\name{package_dfvis}
\alias{dfvis}
\alias{dfvis-package}
\title{About 'dfvis' package}
\description{
A ggplot2 implementation of tabplot
}
\author{
\strong{Maintainer}: Srikanth Komala Sheshachala \email{sri.teach@gmail.com}

}
