% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_column}
\alias{plot_column}
\title{Create plot object corresponding to a single column}
\usage{
plot_column(column_name, qdf)
}
\arguments{
\item{qdf}{dataframe with '.quantile' column}
}
\value{
ggplot object
}
\description{
Plot object depends of the type of variable
}
